+++ package/expo-package/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { FlatList, Image, Platform } from 'react-native';
 
 import NetInfo from '@react-native-community/netinfo';
-import { Video as ExpoVideoPlayer } from 'expo-av';
+
 import * as DocumentPicker from 'expo-document-picker';
 import * as FileSystem from 'expo-file-system';
 import * as Haptics from 'expo-haptics';
@@ -11,6 +11,8 @@ import * as ImagePicker from 'expo-image-picker';
 import * as MediaLibrary from 'expo-media-library';
 import * as Sharing from 'expo-sharing';
 import { registerNativeHandlers } from 'stream-chat-react-native-core';
+
+import ExpoVideoPlayer from './optionalDependencies/Video';
 
 registerNativeHandlers({
   compressImage: async ({ compressImageQuality = 1, uri }) => {
@@ -220,18 +222,18 @@ registerNativeHandlers({
     }
   },
   // eslint-disable-next-line react/display-name
-  Video: ({ onPlaybackStatusUpdate, paused, style, uri, videoRef }) => (
-    <ExpoVideoPlayer
-      onPlaybackStatusUpdate={onPlaybackStatusUpdate}
-      ref={videoRef}
-      resizeMode='contain'
-      shouldPlay={!paused}
-      source={{
-        uri,
-      }}
-      style={[style]}
-    />
-  ),
+  Video: ExpoVideoPlayer ? ({ onPlaybackStatusUpdate, paused, style, uri, videoRef }) => (
+      <ExpoVideoPlayer
+        onPlaybackStatusUpdate={onPlaybackStatusUpdate}
+        ref={videoRef}
+        resizeMode='contain'
+        shouldPlay={!paused}
+        source={{
+          uri,
+        }}
+        style={[style]}
+      />
+    ) : null,
 });
 
 export * from 'stream-chat-react-native-core';