+++ package/src/contexts/suggestionsContext/__tests__/Suggestions.test.js
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { View } from 'react-native';
 
 import { cleanup, render, waitFor } from '@testing-library/react-native';